Comment-out set_wdog_reset in myd_y6ull.c
The call to set_wdog_reset in myd_y6ull.c causes a compilation failure
with "undefined reference to `set_wdog_reset`". It seems that you can
call this function for the MX7 and MX8 archs, but not for IMX6. I have
no idea if commenting the line out will break things, but it is worth
a try.